+++ CHANGELOG.md
@@ -11,9 +11,13 @@ Release notes for the Router Craft CMS plugin.
 - Added a `craft.router` Twig global variable
 - Added `craft.router.params()` to access all named params from the URL
 - Added `craft.router.rawParams()` to access all named params (with raw values) from the URL
+- Added `craft.router.url()` to build a URL out of named routes and optional params
+- Added `craft.router.urlMerge()` to build a URL out of the current route and additional params
 - Added `miranj\router\services\Router`
 - Added `miranj\router\services\Router::getParams()`
 - Added `miranj\router\services\Router::getRawParams()`
+- Added `miranj\router\services\Router::getUrl()`
+- Added `miranj\router\services\Router::getUrlMerge()`
 
 ### Fixed
 - Fixed bug where top level URL rule (with no segments) would get registered twice.